+++ gulpfile.js
@@ -77,7 +77,7 @@ Gulp.registry(new GulpForwardReference());
     'lint-ts',
     'palette',
     'prep-release',
-    'prepare-react-demos',
+    'prepare-react-samples',
     'reset-clone',
     'reset-visual-references',
     'scripts',
