+++ samples/highcharts/react/basic/demo.jsx
@@ -1,6 +1,6 @@
-import React from  'https://esm.sh/react@18.3.1';
-import ReactDOM from 'https://esm.sh/react-dom@18.3.1/client';
-import { Chart, Series } from 'https://esm.sh/gh/highcharts/highcharts-react@v4-dev/index';
+import React from  'react';
+import ReactDOM from 'react-dom';
+import { Chart, Series } from 'highcharts-react-official';
 
 export default function ChartComponent() {
     return (
